Narrowed selector expression for JSON parsing
Parsing the shoutbox message could inadvertently select the wrong element in the DOM if the user had a span element in the prefix/suffix of the member group. Altering the expression narrows down the search and should prevent this, at minimal resource cost (as if script performance was ever a concern when it comes to my handiwork lmao).

An extra option was added, allowing users to enable usage of the username from the forum, complete with prefix and suffixes from the membergroups, both retroactively and going forward.